deprecate `CryptoFileSystemProperties.PROPERTY_MASTERKEY_FILENAME`
This property is only used in one place any longer: To throw FileSystemNeedsMigrationException when attempting to unlock a vault that is lacking a vault config file. This is effectively dead code, since the check is done in the GUI application upfront. We'll keep this behaviour for now but should change the contract eventually